feat: work in progress for arbitrum bridge in swapper (#7344)
* feat: claims ui progression

* chore: break out ConfirmSummary into a separate component

* chore: break out body of trade input into a separate component

* chore: move trade input header into separate component

* chore: make ClaimRow reusable

* feat: wireframing claims

* feat: hide arbitrum bridge claims behind feature flag

* chore: actioned code review feedback
