+++ .env.base
@@ -39,6 +39,7 @@ REACT_APP_FEATURE_RFOX=true
 REACT_APP_FEATURE_RFOX_REWARDS_TX_HISTORY=false
 REACT_APP_FEATURE_ARBITRUM_BRIDGE=false
 REACT_APP_FEATURE_CUSTOM_TOKEN_IMPORT=false
+REACT_APP_FEATURE_ARBITRUM_BRIDGE_CLAIMS=false
 
 # absolute URL prefix
 REACT_APP_ABSOLUTE_URL_PREFIX=https://app.shapeshift.com
