+++ .env.dev
@@ -3,6 +3,7 @@ REACT_APP_FEATURE_RFOX_REWARDS_TX_HISTORY=true
 REACT_APP_FEATURE_ARBITRUM_BRIDGE=true
 REACT_APP_FEATURE_COWSWAP_ARBITRUM=true
 REACT_APP_FEATURE_CUSTOM_TOKEN_IMPORT=true
+REACT_APP_FEATURE_ARBITRUM_BRIDGE_CLAIMS=true
 
 # logging
 REACT_APP_REDUX_WINDOW=false
