+++ src/assets/translations/en/main.json
@@ -766,7 +766,10 @@
       "waitWarning": "Arbitrum’s native bridge is more secure but can take up to a week to claim.",
       "waitCta": "I understand that it will take ~7 days before I can claim funds on Ethereum",
       "claimCta": "I understand that after claiming my funds, I’ll have to send another transaction with a fee"
-    }
+    },
+    "claims": "Claims",
+    "availableClaims": "Available",
+    "pendingClaims": "Pending"
   },
   "trade": {
     "trade": "Trade",
