+++ src/components/ClaimRow/types.ts
@@ -1,0 +1,4 @@
+export enum ClaimStatus {
+  Available = 'Available',
+  Pending = 'Pending',
+}
