+++ src/config.ts
@@ -178,6 +178,7 @@ const validators = {
   REACT_APP_FEATURE_RFOX_REWARDS_TX_HISTORY: bool({ default: false }),
   REACT_APP_FEATURE_RFOX_MOCK_REWARDS_TX_HISTORY: bool({ default: false }),
   REACT_APP_RFOX_REWARDS_MOCK_RUNE_ADDRESS: str(),
+  REACT_APP_FEATURE_ARBITRUM_BRIDGE_CLAIMS: bool({ default: false }),
 }
 
 function reporter<T>({ errors }: envalid.ReporterOptions<T>) {
