+++ src/pages/RFOX/components/RewardsAndClaims/Claims.tsx
@@ -5,6 +5,7 @@ import dayjs from 'dayjs'
 import noop from 'lodash/noop'
 import { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
+import { ClaimStatus } from 'components/ClaimRow/types'
 import { Text } from 'components/Text'
 import { fromBaseUnit } from 'lib/math'
 import { useGetUnstakingRequestsQuery } from 'pages/RFOX/hooks/useGetUnstakingRequestsQuery'
@@ -12,7 +13,6 @@ import { selectAssetById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 import { ClaimRow } from '../Claim/ClaimRow'
-import { ClaimStatus } from '../Claim/types'
 
 type ClaimsProps = {
   headerComponent: JSX.Element
@@ -69,7 +69,7 @@ export const Claims = ({ headerComponent, stakingAssetId, stakingAssetAccountId 
       const isAvailable = currentTimestampMs >= unstakingTimestampMs
       const cooldownDeltaMs = unstakingTimestampMs - currentTimestampMs
       const cooldownPeriodHuman = dayjs(Date.now() + cooldownDeltaMs).fromNow()
-      const status = isAvailable ? ClaimStatus.Available : ClaimStatus.CoolingDown
+      const status = isAvailable ? ClaimStatus.Available : ClaimStatus.Pending
       return (
         <ClaimRow
           stakingAssetId={stakingAssetId}
