+++ src/state/slices/preferencesSlice/preferencesSlice.ts
@@ -62,6 +62,7 @@ export type FeatureFlags = {
   RfoxRewardsTxHistory: boolean
   CustomTokenImport: boolean
   RfoxMockRewardsTxHistory: boolean
+  ArbitrumBridgeClaims: boolean
 }
 
 export type Flag = keyof FeatureFlags
@@ -149,6 +150,7 @@ const initialState: Preferences = {
     RfoxRewardsTxHistory: getConfig().REACT_APP_FEATURE_RFOX_REWARDS_TX_HISTORY,
     CustomTokenImport: getConfig().REACT_APP_FEATURE_CUSTOM_TOKEN_IMPORT,
     RfoxMockRewardsTxHistory: getConfig().REACT_APP_FEATURE_RFOX_MOCK_REWARDS_TX_HISTORY,
+    ArbitrumBridgeClaims: getConfig().REACT_APP_FEATURE_ARBITRUM_BRIDGE_CLAIMS,
   },
   selectedLocale: simpleLocale(),
   balanceThreshold: '0',
